refactor(station): migrate production consumers to dashboard role entrypoints
Rewrite all 49 Station production files importing the ambiguous
@station/dashboard-core root barrel to the explicit role entrypoints
(runtime, state, selectors), one role per import. The frozen
importBoundaries inventories follow the two DashboardRuntime
composition imports onto the runtime entrypoint, and the freeze scan
now collects that specifier alongside the root until the root is
removed.
