+++ station/src/station/view/ToastOverlayView.tsx
@@ -1,14 +1,14 @@
 // Render layer: a bottom-anchored notice that grows upward for actionable errors.
 // Only the header dismiss control routes dismissal; body text stays selectable.
 import { MouseButton, TextAttributes } from "@opentui/core";
-import {
+import { 
   toastBorderColor,
   toastCopyText,
   toastDetail,
   toastOverlayLayout,
   toastTitle,
-  type TuiToastEntry,
-} from "@station/dashboard-core";
+ } from "@station/dashboard-core/selectors";
+import type { TuiToastEntry } from "@station/dashboard-core/state";
 import { useEffect, useState } from "react";
 import {
   toastBorderThemeColor,
