+++ station/src/station/view/sheets/ProjectChoiceSheetView.tsx
@@ -1,10 +1,6 @@
 import type { ProjectId } from "@station/contracts";
-import {
-  bottomSheetContentWidth,
-  selectProjectChooserChoices,
-  type DashboardSnapshotView,
-  type DashboardStateView,
-} from "@station/dashboard-core";
+import { bottomSheetContentWidth, selectProjectChooserChoices } from "@station/dashboard-core/selectors";
+import type { DashboardSnapshotView, DashboardStateView } from "@station/dashboard-core/state";
 import { providerHealthColor, useStationTheme } from "../../../theme/index.js";
 import { BottomSheetFrameView } from "./BottomSheetFrameView.js";
 import { SheetChoiceLine, SheetFooter, SheetLine } from "./parts.js";
